Board as table
The sudoku board has been implemented as a table. And the cell as a text. This has allowed for the reuse of the table class and the text class and the use of the table within a container. Container class has had a rework and the elements (table, button, text and icon) have been reworked to fit the changes. All activities now extend from a base class. The run method has been generalised other than for the WelcomeActivity.